Refactor audio recording to be in a hook.
